+++ src/bin.js
@@ -12,8 +12,8 @@ let collectPips = function (value, list) {
 program
     .version('0.0.1')
     .option('-p, --pretend', 'only output resulting structure without packaging')
-    .option('-g, --gzip', 'use .tar.gz instead of the .tar-format')
     .option('--pip [value]', 'if default files for custom PIPs are used, use this parameter to specify the default', collectPips, {})
+    .option('-d, --destination [value]', 'The path the resulting archive will be saved to (defaults to cwd)', '.')
     .option('-q, --quiet', 'omit any output')
     .parse(process.argv);
 
@@ -23,6 +23,7 @@ program
         gzip: program.gzip,
         pips: program.pip,
         quiet: program.quiet,
-        cwd: process.cwd()
+        cwd: process.cwd(),
+        destination: program.destination,
     }).run()
 })(program)